// packages/fdr-sdk/src/client/generated/api/resources/api/resources/v1/resources/read/resources/endpoint/types/SupportedLanguage.ts
/**
 * This file was auto-generated by Fern from our API Definition.
 */

export type SupportedLanguage =
    | "curl"
    | "python"
    | "javascript"
    | "js"
    | "node"
    | "typescript"
    | "ts"
    | "go"
    | "ruby"
    | "csharp";

export const SupportedLanguage = {
    Curl: "curl",
    Python: "python",
    Javascript: "javascript",
    Js: "js",
    Node: "node",
    Typescript: "typescript",
    Ts: "ts",
    Go: "go",
    Ruby: "ruby",
    Csharp: "csharp",
} as const;